minio: fix minio-Java version to 8.0.3 v2
The following code used in testing was failing:
docker build . -f Dockerfile.mint -t storj/mint

This is because we hard-forked Minio, but it pulled in more recent
version of minio-Java in its "Mint" tests.  So, we just hardcode the
minio-Java to the last working version.

The first fix for this I wasn't paying attention and I couldn't test
because of an unrelated Docker issue.  This version does build locally.

Change-Id: I09532f969ad23f3f09438ae00ec2432d6189a27a