cmd: Forward client IP/host to Authservice
Forward the client IP or host that has originated the request to the
Auth service because it will require it in the near future for
mitigating malicious clients to overwhelm its DB using correct formatted
access key IDs which doesn't resolve to any existing access grant.

Minio stores the request information in the context using a package
function of the logger package named `SetReqInfo`.

`SeqReqInfo` is called by a package function of the cmd package named
`newContext`.

Each endpoint calls `newContext` for setting the request info in the
request context and returns it.

`newContext` uses the function package of the pkg/handlers package named
`GetSourceIP`for retrieving the client IP from the headers (see below)
and falling back to `http.Request.RemoteAddr` when there is not headers
than contain information about the client IP.

`GetSourceIP` calls the function package named `GetSourceIPFromHeaders`
(function in the same package than the former) for figuring out the
client IP from the following headers (in this order) `X-Forwareded-For`,
`X-Real-Ip` and `Forwarded`.