cmd:  fix access grant passing for presigned puts
test_presigned_put_object from minio client functional-tests.sh script
fails, because it tries to use multi-part form data to post credentials.

This is the ONLY handler in the entire MinIO codebase that uses
doesPolicySignatureMatch instead of checkRequestAuthType. Using the
second function means we are unable to populate Access Grant in the
context until we develop authorisation outside of MinIO.

Change-Id: I0e650fe2eb57f8f4962d7d7e40e31371a3fa8ba3