all: change import path
This change makes the code use the storj.io/minio import path instead of
github.com/minio/minio and changes the module path appropriately.

The import path gets changed because it allows us to, e.g.

go install storj.io/gateway-mt@main

and not rely on the replace directive (go install path@version won't run
for modules with replace directives). Given the instructions below, this
change should be easily reversible considering the need to update the
fork.

The import path change was done by using
https://github.com/sirkon/go-imports-rename. The exact steps to
replicate the rename in the root of the repository:

1. go install github.com/sirkon/go-imports-rename@v0.1.2
2. go-imports-rename --save 'github.com/minio/minio/ => storj.io/minio/'
3. go mod edit -module storj.io/minio
4. go install golang.org/x/tools/cmd/goimports@v0.1.7
5. goimports -local storj.io -w .

Change-Id: I802fcd056adc4d8375a6db924bfc8c0d641affd0