cmd: pass request context to GetUser
With the desire in mind to use Minio from Gateway-MT as a library, the
need arises to remove any global state.  The primary cultprit here is
Minio's GlobalIAMSys, which previously needed state to retrieve User
Secret Keys and Access Grants from the Auth Service.

This change passes in the HTTP request context to IAMSys GetUser()
calls, which allow Minio to pull user information from context.  This
means that GlobalIAMSys no longer needs to be stateful.  This change
supports the Gateway-MT changes at
https://review.dev.storj.io/c/storj/gateway-mt/+/5880.

Change-Id: If3e500b82be61d4a78bfe09297ead1ad4d970dac