minio/cmd:  error PutObject if content-length header absent
S3 appears to error if the content-length header is unset.  Minio is
more liberal in that it accepts a put if the actual content length can
be determined.  In order to best approximate s3 and meet the Ceph /
Splunk S3 tests, we're more closely aligning to the S3 behavior.

Updates storj/gateway-mt#150

Change-Id: Ia076af23e25fb088f0b2329b057b7f56f54ef0fd