cmd: stop using ErrMalformedDate, use ErrMissingDateHeader
Amazon appears to always return the same error, and does not return a
specific error for being unable to parse a date.  The S3 docs do not
list the ErrMalformedDate error texts, nor does a Google search indicate
it exists outside of Minio.  The Ceph/Splunk tests require returning the
ErrMissingDateHeader error.

Change-Id: Ida92bf9f21911932c9cd1783f2e32b564e4cacab