cmd: remove global lock checks for CopyObjectHandler
CopyObjectHandler calls checkPutObjectLockAllowed which does some
checks of the request to see if retention is requested. However,
that function doesn't work as expected with our object lock feature
because it relies on globalBucketObjectLockSys in minio, but we
don't use that.

This change removes that functionality so the function is solely
used for validating retention parameters in the request only.

Updates storj/edge#499

Change-Id: I11b9d2e48bf9241a9482c7854dc340f4ee86291f