cmd: fix legal hold not passed through PutObject
this partially reverts fc1f7cd but sets retention and
legal hold options the same as other handlers like
CopyObjectHandler.

additionally, this change fixes parseObjectLockHeaders
to correctly set legalHold variable instead of throwing
it away if only legal hold was requested, but not legal
hold + retention. This was a regression introduced in
00eca4d.

Change-Id: I93b6d5bcdc79287b25636b1ca52fb65f18e9200f