cmd: support GetObjectAttributes
Support GetObjectAttributes for the following attributes:

- ETag
- StorageClass
- ObjectSize

If ObjectParts or Checksum is requested, we allow them in the request
but omit the fields from the response. It seemed better to do this
for client compatibility instead of failing these requests.

StorageClass has a special case of always returning "STANDARD" since
we don't have a concept of classes, nor do we store the requested
class when uploading an object.

Change-Id: I6c2d62ca8df9373122306dca08b80cc99babf88c