cmd: restructure response of DeleteObjects object layer method
This change retructures the response of the object layer interface's
DeleteObjects method.

Previously, DeleteObjects implementations were required to return a
slice of all objects for which deletion was attempted and a
corresponding slice of errors that were encountered as a result.

Now, implementations are required to return a slice of deletion
successes, a slice of deletion errors, and, if applicable, the error
that prevented the processing of the request.

References storj/edge#521

Change-Id: I4f4b8e134a4154fe3d53e93915217e4b0e715564