cmd: revert removal of auth logic in several request handlers
This change reintroduces a method call responsible for request auth
that was removed from several request handlers in the process of
reworking them to account for object API updates.

The method both validates the request signature and checks whether the
requested action is allowed. While the access control logic isn't
necessary for our use case since it's handled by the satellite,
signature validation is since it ensures that the request isn't
malformed.

The affected operations are as follows:

- PutObjectLockConfiguration
- GetObjectLockConfiguration
- PutBucketTagging
- GetBucketTagging
- DeleteBucketTagging

Change-Id: Ic8f728f3d2910e6d21964f98047549fb2fe84ce4