cmd: fix nil pointer dereference of ObjectAPIHandlers' CacheObjectLayer
This change fixes an issue where object API handlers would attempt to
dereference a nil instance of a cache object API layer.

ObjectAPIHandlers can be constructed with a function that provides a
CacheObjectLayer implementation. Previously, the wrapper around this
provider would report that a CacheObjectLayer was available even when
the provider returned nil. This caused ObjectAPIHandlers' HTTP handlers
to attempt to use this value, causing a nil pointer dereference.

This issue has been fixed by properly handling the case where the
CacheObjectLayer provider returns a nil value.

Change-Id: Ibf20d316e90bf17a21418b97d0886926dac27b85